Issue #496  Replace or augment Tcl code with Python equivalents
Factored the state poller and its factory out into rdo_utils
so that polling can be common across controllers.
